`CenteredContent` moved to `box` subdirectory
