+++ apps/emergence/app/page.tsx
@@ -1,5 +1,5 @@
 import { type ReactElement } from 'react'
-import { CenteredContent } from '@repo/mui/components/CenteredContent'
+import { CenteredContent } from '@repo/mui/components/box/CenteredContent'
 import { Copyright } from '@repo/mui/components/Copyright'
 import { PageContainer } from '@repo/mui/components/PageContainer'
 import { ProTip } from '@repo/mui/components/ProTip'
