+++ packages/mui/src/components/box/CenteredContent.tsx
